finished task 3
I monitor the input event on the form, and each time I write an object with the email and message fields to the local storage, which stores the current values of the form fields. The key for the storage is the string "feedback-form-state".
When the page loads, I check the state of the storage, and if there is any saved data there, I fill in the form fields with it. Otherwise, the fields will be empty.
During the form summit, I clear the storage and form fields, and display an object with the email, message fields and their current values in the console.
I use the lodash.throttle library to ensure that the storage is updated no more than once every 500 milliseconds.